Revert "Fix Process ID mismatch"
This reverts commit 9d1cd63a3dbefb1c68cd4524b86787d29d9d5089.

+++ include/proc.h
@@ -2,6 +2,6 @@
 #define NOAH_PROC_H
 
 int platform_clone_process(unsigned long clone_flags, unsigned long newsp, gaddr_t parent_tid, gaddr_t child_tid, gaddr_t tls);
-int platform_restore_proc(uint64_t pid);
+int platform_restore_proc(unsigned pid);
 
 #endif
